# docs/specifications/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Dog Walking Management API
  description: |
    The Stardogwalker management API — Cardiff, South Wales.

    This API is designed **API-first** to support multiple interaction
    channels: web app, native mobile app, conversational chat interfaces
    (e.g. WhatsApp bot), and agentic AI applications. All business logic
    lives here; no channel-specific logic is embedded in the API.

    **Design principles:**
    - Stateless — every request authenticates via JWT bearer token
    - Fine-grained actions — discrete endpoints for each state transition
      (e.g. `/accept`, `/decline`) to support unambiguous intent from any
      client type including AI agents
    - Consistent response envelopes — list responses always include
      `data[]` and `pagination`; errors always return `code`, `message`,
      and optional `details`
    - Versioned — breaking changes will be introduced under `/v2` without
      removing `/v1` until all clients have migrated

    Covers: interest registration (public), authentication, dog owner and
    walker management, dog profiles, walk requests, walk scheduling and
    execution, real-time walk updates, and invoicing.
  version: 1.0.0
  contact:
    name: Dog Walking API Support
    email: support@dogwalking.example.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/v1
    description: Local development server
  - url: https://api.dogwalking.example.com/v1
    description: Production server
...
security:
  - bearerAuth: []
...
paths:
...
  /walkers:
    get:
      tags:
        - Walkers
      summary: List dog walkers
      description: Retrieve a paginated list of dog walkers.
      operationId: listWalkers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: A list of dog walkers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkerList'
        '401':
          $ref: '#/components/responses/Unauthorised'

    post:
      tags:
        - Walkers
      summary: Create a dog walker
      description: Create a new dog walker profile.
      operationId: createWalker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkerRequest'
      responses:
        '201':
          description: Dog walker created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walker'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
  /walkers/{walkerId}:
    parameters:
      - $ref: '#/components/parameters/WalkerId'
    get:
      tags:
        - Walkers
      summary: Get a dog walker
      description: Retrieve details of a specific dog walker.
      operationId: getWalker
      responses:
        '200':
          description: Dog walker details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walker'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Walkers
      summary: Update a dog walker
      description: Update the profile of a specific dog walker.
      operationId: updateWalker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkerRequest'
      responses:
        '200':
          description: Dog walker updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walker'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Walkers
      summary: Delete a dog walker
      description: Delete a dog walker profile.
      operationId: deleteWalker
      responses:
        '204':
          description: Dog walker deleted
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Dogs
  # ---------------------------------------------------------------------------
...
  /walk-requests:
    get:
      tags:
        - Walk Requests
      summary: List walk requests
      description: |
        Retrieve a paginated list of walk requests. Dog owners see their own
        requests; dog walkers see requests assigned to them or available requests.
      operationId: listWalkRequests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter requests by status
          schema:
            $ref: '#/components/schemas/WalkRequestStatus'
      responses:
        '200':
          description: A list of walk requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkRequestList'
        '401':
          $ref: '#/components/responses/Unauthorised'

    post:
      tags:
        - Walk Requests
      summary: Create a walk request
      description: A dog owner submits a new walk request.
      operationId: createWalkRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkRequestCreate'
      responses:
        '201':
          description: Walk request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
  /walk-requests/{requestId}:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      tags:
        - Walk Requests
      summary: Get a walk request
      description: Retrieve details of a specific walk request.
      operationId: getWalkRequest
      responses:
        '200':
          description: Walk request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Walk Requests
      summary: Update a walk request
      description: Update a pending walk request (e.g. change date/time or dogs).
      operationId: updateWalkRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkRequestCreate'
      responses:
        '200':
          description: Walk request updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Walk Requests
      summary: Cancel a walk request
      description: Cancel and remove a walk request.
      operationId: cancelWalkRequest
      responses:
        '204':
          description: Walk request cancelled
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /walk-requests/{requestId}/accept:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    post:
      tags:
        - Walk Requests
      summary: Accept a walk request
      description: A dog walker accepts a walk request, creating a scheduled walk.
      operationId: acceptWalkRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkRequestAccept'
      responses:
        '200':
          description: Walk request accepted and walk scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /walk-requests/{requestId}/decline:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    post:
      tags:
        - Walk Requests
      summary: Decline a walk request
      description: A dog walker declines a walk request.
      operationId: declineWalkRequest
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkRequestDecline'
      responses:
        '200':
          description: Walk request declined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Walks
  # ---------------------------------------------------------------------------
  /walks:
    get:
      tags:
        - Walks
      summary: List walks
      description: |
        Retrieve a paginated list of scheduled or completed walks. Dog owners
        see walks for their dogs; walkers see their assigned walks.
      operationId: listWalks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter walks by status
          schema:
            $ref: '#/components/schemas/WalkStatus'
        - name: date
          in: query
          description: Filter walks by date (YYYY-MM-DD)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: A list of walks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkList'
        '401':
          $ref: '#/components/responses/Unauthorised'
  /walks/{walkId}:
    parameters:
      - $ref: '#/components/parameters/WalkId'
    get:
      tags:
        - Walks
      summary: Get a walk
      description: Retrieve details of a specific walk.
      operationId: getWalk
      responses:
        '200':
          description: Walk details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Walks
      summary: Update a walk
      description: Update details of a scheduled walk (e.g. start time, route notes).
      operationId: updateWalk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkEdit'
      responses:
        '200':
          description: Walk updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    WalkerId:
      name: walkerId
      in: path
      required: true
      description: Unique identifier of the dog walker
      schema:
        type: string
        format: uuid
...
    RequestId:
      name: requestId
      in: path
      required: true
      description: Unique identifier of the walk request
      schema:
        type: string
        format: uuid
    WalkId:
      name: walkId
      in: path
      required: true
      description: Unique identifier of the walk
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Number of results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validationError:
              summary: Validation error
              value:
                code: VALIDATION_ERROR
                message: One or more fields failed validation.
                details:
                  - field: email
                    issue: Must be a valid email address.
                  - field: durationMinutes
                    issue: Must be between 15 and 240.
            noMatchingRate:
              summary: No matching rate found
              value:
                code: NO_MATCHING_RATE
                message: No rate found for this walk type, dog count, and duration. Please update the walker's rate card.
            invalidStateTransition:
              summary: Invalid state transition
              value:
                code: INVALID_STATE_TRANSITION
                message: This action cannot be performed on a resource in its current status.
            schedulingConflict:
              summary: Scheduling conflict
              value:
                code: SCHEDULING_CONFLICT
                message: The walker already has an overlapping walk scheduled for this time.
    Unauthorised:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingToken:
              summary: Missing bearer token
              value:
                code: AUTHENTICATION_REQUIRED
                message: A valid bearer token is required to access this resource.
            invalidCredentials:
              summary: Invalid login credentials
              value:
                code: INVALID_CREDENTIALS
                message: The email or password provided is incorrect.
            tokenExpired:
              summary: Expired token
              value:
                code: TOKEN_EXPIRED
                message: The access token has expired. Please refresh your session.
    Forbidden:
      description: Authenticated user does not have permission to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            insufficientRole:
              summary: Insufficient role
              value:
                code: FORBIDDEN
                message: Your role does not have permission to perform this action.
            notOwnResource:
              summary: Not own resource
              value:
                code: FORBIDDEN
                message: You can only access your own records.
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            resourceNotFound:
              summary: Resource not found
              value:
                code: RESOURCE_NOT_FOUND
                message: The requested resource could not be found.
    Conflict:
      description: The request conflicts with the current state of the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            duplicateEmail:
              summary: Duplicate email
              value:
                code: DUPLICATE_EMAIL
                message: A user with this email address already exists.
            alreadyAccepted:
              summary: Already accepted
              value:
                code: ALREADY_PROCESSED
                message: This request has already been accepted or declined.

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    WalkerRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - phone
      properties:
        firstName:
          type: string
          example: Bob
        lastName:
          type: string
          example: Jones
        email:
          type: string
          format: email
          example: bob.jones@example.com
        phone:
          type: string
          example: +44 7700 900456
        address:
          $ref: '#/components/schemas/Address'
        bio:
          type: string
          description: Short biography or description of the walker
          example: Experienced dog walker with 5 years looking after all breeds.
        walkRates:
          type: array
          description: |
            Rate card for this walker. The rate applied to a walk is determined
            by matching walkType, numberOfDogs, and durationMinutes against
            this list.
          items:
            $ref: '#/components/schemas/WalkRate'
    Walker:
      allOf:
        - $ref: '#/components/schemas/WalkerRequest'
        - type: object
          required:
            - id
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              format: uuid
              example: 7b1e4d2a-1234-5678-abcd-ef0123456789
            createdAt:
              type: string
              format: date-time
              example: '2025-01-10T08:00:00Z'
            updatedAt:
              type: string
              format: date-time
              example: '2025-01-10T08:00:00Z'
    WalkerList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Walker'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # --- Dogs -----------------------------------------------------------------
...
    WalkRequestStatus:
      type: string
      enum:
        - pending
        - accepted
        - declined
        - cancelled
      example: pending
...
    WalkRequestCreate:
      type: object
      required:
        - ownerId
        - dogIds
        - requestedDate
        - requestedStartTime
        - durationMinutes
        - walkType
      properties:
        ownerId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        dogIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
          example:
            - c1a2b3d4-5678-90ab-cdef-1234567890ab
        preferredWalkerId:
          type: string
          format: uuid
          description: Optional preferred walker
          example: 7b1e4d2a-1234-5678-abcd-ef0123456789
        requestedDate:
          type: string
          format: date
          example: '2025-04-10'
        requestedStartTime:
          type: string
          format: time
          description: Requested start time in HH:MM:SS format (RFC 3339)
          example: '09:00:00Z'
        durationMinutes:
          type: integer
          minimum: 15
          maximum: 240
          description: Requested walk duration in minutes
          example: 60
        walkType:
          $ref: '#/components/schemas/WalkType'
        notes:
          type: string
          description: Special instructions for the walker
          example: Please use the harness hanging by the front door.
    WalkRequestAccept:
      type: object
      required:
        - walkerId
        - confirmedDate
        - confirmedStartTime
      properties:
        walkerId:
          type: string
          format: uuid
          example: 7b1e4d2a-1234-5678-abcd-ef0123456789
        confirmedDate:
          type: string
          format: date
          example: '2025-04-10'
        confirmedStartTime:
          type: string
          format: time
          example: '09:00:00Z'
        notes:
          type: string
          example: I'll be there at 9am sharp!
    WalkRequestDecline:
      type: object
      properties:
        reason:
          type: string
          description: Reason for declining the request
          example: I am fully booked on that date.
    WalkRequest:
      type: object
      required:
        - id
        - ownerId
        - dogIds
        - requestedDate
        - requestedStartTime
        - durationMinutes
        - walkType
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: aa11bb22-cc33-dd44-ee55-ff6677889900
        ownerId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        dogIds:
          type: array
          items:
            type: string
            format: uuid
          example:
            - c1a2b3d4-5678-90ab-cdef-1234567890ab
        preferredWalkerId:
          type: string
          format: uuid
          nullable: true
          example: 7b1e4d2a-1234-5678-abcd-ef0123456789
        requestedDate:
          type: string
          format: date
          example: '2025-04-10'
        requestedStartTime:
          type: string
          format: time
          example: '09:00:00Z'
        durationMinutes:
          type: integer
          example: 60
        walkType:
          $ref: '#/components/schemas/WalkType'
        notes:
          type: string
          description: Special instructions for the walker
          example: Please use the harness hanging by the front door.
        status:
          $ref: '#/components/schemas/WalkRequestStatus'
        declineReason:
          type: string
          nullable: true
          example: I am fully booked on that date.
        walkId:
          type: string
          format: uuid
          nullable: true
          description: Set when a walk is created from this request
          example: null
        createdAt:
          type: string
          format: date-time
          example: '2025-04-01T14:00:00Z'
        recurringWalkId:
          type: string
          format: uuid
          nullable: true
          description: Set when this request was generated from a recurring walk schedule
          example: null
        updatedAt:
          type: string
          format: date-time
          example: '2025-04-01T14:00:00Z'
    WalkRequestList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/WalkRequest'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # --- Walks ----------------------------------------------------------------
    WalkStatus:
      type: string
      enum:
        - scheduled
        - in_progress
        - completed
        - cancelled
      example: scheduled
...